+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # 2.x branch
 ## 2.12 branch
 ### 2.12.2
+* added `HtmlHelper::iconClass()` method;
 * `Youtube::getId()` method takes short url with the duration (eg.
-    `http://youtu.be/bL_CJKq9rIw?t=5s`).
+    `http://youtu.be/bL_CJKq9rIw?t=5s`);
+* tests that require a network connection have been marked with the
+    `requireNetwork` group.
 
 ### 2.12.1
 * methods that have been deprecated with CakePHP 3.4 have been replaced.